Moves util functions to util file
